requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
elevenlabs>=0.2.0
TTS>=0.22.0
requests>=2.31.0
Pillow>=10.0.0
pyttsx3>=2.90 